+++ src/hm/videostore/renting/api/Rental.java
@@ -1,10 +1,11 @@
 package hm.videostore.renting.api;
 
 import hm.videostore.statement.api.StatementItem;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface Rental {
     StatementItem returnOn(LocalDate returnDate);
 
-    double getPriceReturnedOn(LocalDate returnDate);
+    BigDecimal getPriceReturnedOn(LocalDate returnDate);
 }
