+++ src/hm/videostore/statement/api/StatementItem.java
@@ -1,5 +1,7 @@
 package hm.videostore.statement.api;
 
+import java.math.BigDecimal;
+
 public interface StatementItem {
-    double getPrice();
+    BigDecimal getPrice();
 }
